requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# AI/ML Dependencies
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.3
librosa==0.10.1
scikit-learn==1.3.2

# Experiment Tracking
clearml==1.13.2

# Data Processing
pandas==2.1.3
pyyaml==6.0.1
soundfile==0.12.1

# Audio Processing
pyaudio==0.2.11
wave==0.0.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Visualization (for development/debugging)
matplotlib==3.8.2
seaborn==0.13.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4